centralized styles for profile page user data components
